+++ src/Phormium/NullLogger.php
@@ -7,10 +7,27 @@ namespace Phormium;
  */
 class NullLogger
 {
-    public function trace() { }
-    public function debug() { }
-    public function info() { }
-    public function warn() { }
-    public function error() { }
-    public function fatal() { }
+    public function trace()
+    {
+    }
+
+    public function debug()
+    {
+    }
+
+    public function info()
+    {
+    }
+
+    public function warn()
+    {
+    }
+
+    public function error()
+    {
+    }
+
+    public function fatal()
+    {
+    }
 }