+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Phormium Changelog
 ==================
 
+0.3 / 2013-06-14
+----------------
+
+* Added QuerySet::valuesFlat()
+* Added optional parameter $allowEmpty to QuerySet::single()
+
 0.2 / 2013-05-10
 ----------------